File clean up complete
