+++ controllers/api/index.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const userRoutes = require('./userRoutes');
-const registrationRoutes = require('./registrationRoutes+');
+const registrationRoutes = require('./registrationRoutes');
 
 router.use('/users', userRoutes);
 router.use('/registration', registrationRoutes);
