+++ package.json
@@ -5,7 +5,7 @@
   "main": "server+.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server+.js",
+    "start": "node server.js",
     "seed": "node seeds/index.js"
   },
   "author": "Pete Smith, Pranith Gunda, David Rauf, Stephanie Carr, Joseph Ruff",
