+++ public/index.html
@@ -23,7 +23,7 @@
           <!-- Login -->
           <ul class="navbar-nav">
             <li class="nav-item">
-              <a class="nav-link" href="user-registration+.html">Register</a>
+              <a class="nav-link" href="user-registration.html">Register</a>
             </li>
             <li class="nav-item">
               <a class="nav-link" href="login.html">Login</a>
