+++ public/login.html
@@ -27,28 +27,6 @@
     </header>
     <main class="d-flex align-center justify-center">
         <div class="col-6 col-sm-8">
-<<<<<<< HEAD
-            <h3 class="display-6 text-center">Log In</h3>
-            <div class="row justify-content-center">
-                <div class="card rounded bg-light p-3">
-                    <div class="card-body">
-                        <form>
-                            <div class="form-group">
-                                <label for="email">Email Address</label>
-                                <input type="email" class="form-control" id="email" placeholder="Enter email">
-                            </div>
-                            <div class="form-group">
-                                <label for="password">Password</label>
-                                <input type="password" class="form-control" id="password" placeholder="Password">
-                            </div>
-                            <button type="submit" class="btn btn-login">Log In</button>
-                        </form>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </main>
-=======
             <h4 class="align-center justify-center">Sign In</h4>
             <div class="card rounded bg-light p-3">
                 <div class="card-body">
@@ -69,7 +47,6 @@
         </div>
     </div>
     <script src="/js/login.js"></script>
->>>>>>> a128f7d790edd387b2f42d9f65dea3a42bc6bcfb
 </body>
 
 </html>
