--- public/user-registration+.html
@@ -1,87 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <!-- bootstrap -->
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous" />
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css" />
-  <!-- Google Fonts -->
-  <style>
-    @import url('https://fonts.googleapis.com/css2?family=Comforter+Brush&family=Open+Sans:ital,wdth,wght@0,75..100,300..800;1,75..100,300..800&family=Patrick+Hand+SC&display=swap');
-    </style>
-  <link rel="stylesheet" href="./assets/css/style.css" />
-  <title>User Registration</title>
-</head>
-
-<body>
-  <!-- Sticky Header -->
-  <header id="header">
-    <nav class="navbar navbar-expand-sm sticky-top">
-      <div class="container-fluid">
-        <a class="navbar-brand" href="index.html">TutorHub </a>
-        <!-- Login -->
-        <ul class="navbar-nav">
-          <li class="nav-item">
-            <a class="nav-link" href="user-registration.html">Register</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="login.html">Login</a>
-          </li>
-        </ul>
-      </div>
-    </nav>
-  </header>
- <form id="registration-form" action="/api/registration" method="POST">
-  <main class="row justify-content-center">
-    <div class="card col-10">
-      <h2 class="mb-4">New Account Registration</h2>
-      <h5>Account Type</h5>
-      <div class="form-check">
-        <input class="form-check-input" type="radio" name="userRole" value="tutor" id="tutor-radio">
-        <label class="form-check-label" for="tutor-radio">
-          Tutor Account
-        </label>
-      </div>
-      <div class="form-check">
-        <input class="form-check-input" type="radio" name="userRole" value="student" id="student-radio">
-        <label class="form-check-label mb-3" for="student-radio">
-          Student Account
-        </label>
-      </div>
-      <h5>User Information</h5>
-      <div class="row">
-        <div class="col-4">
-          <input name="firstName" type="text" class="form-control" placeholder="First name" aria-label="First name">
-        </div>
-        <div class="col-5 mb-3">
-          <input name="lastName" type="text" class="form-control" placeholder="Last name" aria-label="Last name">
-        </div>
-      </div>
-      <label for="email-input" class="form-label">Email address <i>(will be your username)</i></label>
-      <input name="email" type="email" class="form-control mb-3" id="email-input" placeholder="name@example.com">
-      <label for="zip code" class="form-label">Zipcode </label>
-      <input name="zipcode" type="text" class="form-control mb-3" id="zip-input" placeholder="34695">
-      <label for="inputPassword" class="form-label">Choose a Password</label>
-      <input name="password" type="password" class="form-control" aria-describedby="passwordHelpBlock">
-      <div id="passwordHelpBlock" class="form-text mb-3">
-        Your password must be 8-20 characters long, contain letters and numbers, and must not contain spaces, special characters, or emoji.
-      </div>
-
-      <button type="submit" id="submit" value="Submit" class="btn mb-3">Submit</button>
-    </div>
-  </main>
-</form>
-      </div>
-    </main>
-  </form>
-  
-  <!-- Including Bootstrap JS -->
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
-    crossorigin="anonymous"></script>
-</body>
-
-</html>
