+++ render.yaml
@@ -14,6 +14,6 @@ services:
     sync: false
   region: oregon
   buildCommand: npm install
-  startCommand: node server+.js
+  startCommand: node server.js
   autoDeploy: false
 version: "1"
